+++ sql/setupcraft_schema.sql
@@ -35,6 +35,7 @@ CREATE TABLE IF NOT EXISTS files (
     source_path TEXT NOT NULL,
     destination_path TEXT NOT NULL,
     install_scope TEXT DEFAULT '',
+    inno_flags TEXT DEFAULT '',
     FOREIGN KEY(project_id) REFERENCES projects(id) ON DELETE CASCADE
 );